requirements.txt
asgiref==3.3.1
django==3.1.7
pytz==2021.1
sqlparse==0.4.1
django-crispy-forms==1.11.2
django-environ==0.9.0
flake8==6.0.0
django-channels
channels_redis

